Code cleanup and pch .
