+++ 3D_Object_Viewer.vcxproj
@@ -156,18 +156,18 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
     <ClCompile Include="src\Rendering\SceneRenderer.cpp" />
-    <ClCompile Include="src\Rendering\OpenGLBuffer.cpp" />
+    <ClCompile Include="src\gpu_side\OpenGLBuffer.cpp" />
     <ClCompile Include="src\Scene\Skybox.cpp" />
     <ClCompile Include="src\Scene\Model.cpp" />
     <ClCompile Include="src\Rendering\Renderer.cpp" />
     <ClCompile Include="src\Scene\Material.cpp" />
-    <ClCompile Include="src\InputHandler.cpp" />
-    <ClCompile Include="src\Rendering\Framebuffer.cpp" />
-    <ClCompile Include="src\Application.cpp" />
+    <ClCompile Include="src\Controls\InputHandler.cpp" />
+    <ClCompile Include="src\gpu_side\Framebuffer.cpp" />
+    <ClCompile Include="src\Core\Application.cpp" />
     <ClCompile Include="src\Scene\Mesh.cpp" />
     <ClCompile Include="src\Helpers\ModelLoader.cpp" />
     <ClCompile Include="src\Rendering\RenderingCoordinator.cpp" />
-    <ClCompile Include="src\Rendering\ShaderClass.cpp" />
+    <ClCompile Include="src\gpu_side\ShaderClass.cpp" />
     <ClCompile Include="src\Scene\Camera.cpp" />
     <ClCompile Include="Dependencies\imgui\imgui\backends\imgui_impl_glfw.cpp" />
     <ClCompile Include="Dependencies\imgui\imgui\backends\imgui_impl_opengl3.cpp" />
@@ -177,11 +177,11 @@
     <ClCompile Include="Dependencies\imgui\imgui\imgui_tables.cpp" />
     <ClCompile Include="Dependencies\imgui\imgui\imgui_widgets.cpp" />
     <ClCompile Include="src\3D_Object_Viewer.cpp" />
-    <ClCompile Include="src\VertexAttributeObject.cpp" />
-    <ClCompile Include="src\WindowHandler.cpp" />
-    <ClCompile Include="src\Texture.cpp" />
+    <ClCompile Include="src\gpu_side\VertexAttributeObject.cpp" />
+    <ClCompile Include="src\Core\WindowHandler.cpp" />
+    <ClCompile Include="src\gpu_side\Texture.cpp" />
     <ClCompile Include="src\UI\UILayer.cpp" />
-    <ClCompile Include="src\UniformBuffer.cpp" />
+    <ClCompile Include="src\gpu_side\UniformBuffer.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="Dependencies\glew\glew-2.2.0\build\vc15\glew_static.vcxproj">
@@ -195,8 +195,8 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="src\AppEvent.h" />
-    <ClInclude Include="src\CommonAppData.h" />
+    <ClInclude Include="src\Core\AppEvent.h" />
+    <ClInclude Include="src\Core\CommonAppData.h" />
     <ClInclude Include="src\Controls\AppKeyBindings.h" />
     <ClInclude Include="src\Controls\KeyboardKeys.h" />
     <ClInclude Include="src\Controls\MouseKeys.h" />
@@ -204,24 +204,24 @@
     <ClInclude Include="src\Helpers\UnitTransformer.h" />
     <ClInclude Include="src\Logging\App_Logger.h" />
     <ClInclude Include="src\pch\pch.h" />
-    <ClInclude Include="src\RenderingMediator.h" />
+    <ClInclude Include="src\Rendering\RenderingMediator.h" />
     <ClInclude Include="src\Rendering\SceneRenderer.h" />
-    <ClInclude Include="src\Rendering\OpenGLBuffer.h" />
+    <ClInclude Include="src\gpu_side\OpenGLBuffer.h" />
     <ClInclude Include="src\Scene\Skybox.h" />
     <ClInclude Include="src\Scene\Model.h" />
     <ClInclude Include="src\Rendering\Renderer.h" />
     <ClInclude Include="src\Scene\Material.h" />
-    <ClInclude Include="src\InputHandler.h" />
-    <ClInclude Include="src\Rendering\Framebuffer.h" />
+    <ClInclude Include="src\Controls\InputHandler.h" />
+    <ClInclude Include="src\gpu_side\Framebuffer.h" />
     <ClInclude Include="Shaders\ShaderPath.h" />
-    <ClInclude Include="src\Application.h" />
+    <ClInclude Include="src\Core\Application.h" />
     <ClInclude Include="src\Helpers\DialogWrapper.h" />
-    <ClInclude Include="src\Events.h" />
+    <ClInclude Include="src\Core\Events.h" />
     <ClInclude Include="src\Scene\Mesh.h" />
-    <ClInclude Include="src\MeshGeneratingMethods.h" />
+    <ClInclude Include="src\Helpers\MeshGeneratingMethods.h" />
     <ClInclude Include="src\Helpers\ModelLoader.h" />
     <ClInclude Include="src\Rendering\RenderingCoordinator.h" />
-    <ClInclude Include="src\Rendering\ShaderClass.h" />
+    <ClInclude Include="src\gpu_side\ShaderClass.h" />
     <ClInclude Include="src\Scene\Camera.h" />
     <ClInclude Include="Dependencies\imgui\imgui\backends\imgui_impl_glfw.h" />
     <ClInclude Include="Dependencies\imgui\imgui\backends\imgui_impl_opengl3.h" />
@@ -231,11 +231,11 @@
     <ClInclude Include="Dependencies\imgui\imgui\imstb_rectpack.h" />
     <ClInclude Include="Dependencies\imgui\imgui\imstb_textedit.h" />
     <ClInclude Include="Dependencies\imgui\imgui\imstb_truetype.h" />
-    <ClInclude Include="src\VertexAttributeObject.h" />
-    <ClInclude Include="src\WindowHandler.h" />
-    <ClInclude Include="src\Texture.h" />
+    <ClInclude Include="src\gpu_side\VertexAttributeObject.h" />
+    <ClInclude Include="src\Core\WindowHandler.h" />
+    <ClInclude Include="src\gpu_side\Texture.h" />
     <ClInclude Include="src\UI\UILayer.h" />
-    <ClInclude Include="src\UniformBuffer.h" />
+    <ClInclude Include="src\gpu_side\UniformBuffer.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Shaders\ClearColorMeshShader.glsl" />
