--- src/CommonAppData.h
@@ -1,18 +1,0 @@
-#pragma once
-namespace OBJ_Viewer {
-	struct Size2D
-	{
-		int width;
-		int height;
-	};
-	struct Position2D
-	{
-		double x;
-		double y;
-	};
-
-	/*using vec2 = glm::vec2;
-	using vec3 = glm::vec3;
-	using mat4 = glm::mat4;
-	using mat3 = glm::mat3;*/
-}
