+++ src/Core/WindowHandler.cpp
