+++ src/Core/WindowHandler.h
