+++ src/Rendering/RenderingMediator.h
