+++ src/Scene/Mesh.h
@@ -1,6 +1,6 @@
 #pragma once
-#include "VertexAttributeObject.h"
-#include"Material.h"
+#include "gpu_side/VertexAttributeObject.h"
+#include "Material.h"
 #include "pch.h"
 
 namespace OBJ_Viewer
